pyproject.toml
[project]
name = "news-curator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "agno==2.4.8",
    "ddgs==9.10.0",
    "duckduckgo-search==8.1.1",
    "fastapi==0.129.0",
    "openai==2.17.0",
    "python-dotenv>=1.2.2",
]
